+++ glutin/src/api/glx/context.rs
@@ -104,7 +104,7 @@ impl Display {
                 return Err(ErrorKind::NotSupported(
                     "extension to create ES context with glx is not present.",
                 )
-                .into())
+                .into());
             },
         };
 
@@ -442,11 +442,7 @@ impl ContextInner {
     }
 
     fn context_api(&self) -> ContextApi {
-        if self.is_gles {
-            ContextApi::Gles(None)
-        } else {
-            ContextApi::OpenGl(None)
-        }
+        if self.is_gles { ContextApi::Gles(None) } else { ContextApi::OpenGl(None) }
     }
 }