+++ glutin/src/lib.rs
@@ -16,7 +16,6 @@
 
 #![deny(rust_2018_idioms)]
 #![deny(rustdoc::broken_intra_doc_links)]
-#![deny(unsafe_op_in_unsafe_fn)]
 #![deny(improper_ctypes, improper_ctypes_definitions)]
 #![deny(clippy::all)]
 #![deny(missing_debug_implementations)]