+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Unreleased
 
 - **Breaking:** Added `make_current_surfaceless(self)` for `{Possibly,Not}CurrentGlContext`.
+- Updated `objc2` frameworks dependency to `v0.3.2`.
+- Use `objc2-open-gl` instead of `cgl` dependency.
+- Bump MSRV from `1.71` to `1.85`.
 
 # Version 0.32.3